cleaned up code, moved into classes
