+++ app/src/main/java/com/detroitlabs/detroitvolunteers/client/SearchOpportunitiesCallBack.java
@@ -1,0 +1,7 @@
+package com.detroitlabs.detroitvolunteers.client;
+
+import com.detroitlabs.detroitvolunteers.client.models.OpportunitiesResponse;
+
+public interface SearchOpportunitiesCallBack {
+    void onSuccess(OpportunitiesResponse response);
+}
